feat(influx_shard): add tool to import shard
Do not use this tool on production systems. Its intended use is quickly
importing an entire shard into an InfluxDB 1.7+ OSS insance for
debugging query or other raw data issues.

This tool can only be run while influxd is offline. It works by copying
the source shard (.tsm and other files) directly into the data directory
and re-writing meta.db to inject the shard.

This is significantly faster than exporting to LP and then importing
that LP. It takes roughly as long as a file copy of the entire shard
directory.