fix(endpoints, rules, checks): route user to alerting index
These are the final tests and rule fix for this issue.

The checks is special in how it was fixed.

But endpoints and rules have the exact same fix on them

All the tests are running smoothly

EsLint was run so this branch should no longer fail from that as well.