feat(influxd): add ability to generate org/bucket
This commit adds a new tool: `influxd generate meta`, which can be used
to create meta data such as organisations and buckets. Initially the
tool is capable of adding a single bucket and organisation.

This tool differs from `influx org create` in that you can specify exact
org and bucket ids, which is useful when you want to create the meta
data associated with existing TSM data.